+++ opentelemetry-api/src/opentelemetry/propagators/__init__.py
@@ -82,16 +82,16 @@ logger = getLogger(__name__)
 
 
 def extract(
-    get_from_carrier: textmap.Getter[textmap.TextMapPropagatorT],
+    get_from_carrier: textmap.Getter,
     carrier: textmap.TextMapPropagatorT,
     context: typing.Optional[Context] = None,
 ) -> Context:
     """ Uses the configured propagator to extract a Context from the carrier.
 
     Args:
-        get_from_carrier: a function that can retrieve zero
-            or more values from the carrier. In the case that
-            the value does not exist, return an empty list.
+        get_from_carrier: an object which contains a get function that can retrieve zero
+            or more values from the carrier and a keys function that can get all the keys
+            from carrier.
         carrier: and object which contains values that are
             used to construct a Context. This object
             must be paired with an appropriate get_from_carrier