+++ opentelemetry-sdk/CHANGELOG.md
@@ -10,6 +10,8 @@
   ([#1105](https://github.com/open-telemetry/opentelemetry-python/pull/1120))
 - Allow for Custom Trace and Span IDs Generation - `IdsGenerator` for TracerProvider
   ([#1153](https://github.com/open-telemetry/opentelemetry-python/pull/1153))
+- Add keys method to TextMap propagator Getter
+  ([#1084](https://github.com/open-telemetry/opentelemetry-python/issues/1084))
 
 ## Version 0.13b0